refactor(backend): fix number type and comment

+++ backend/src/main/java/io/github/changmin6362/multistore/feature/authorization/controller/PermissionController.java
@@ -5,7 +5,7 @@ import io.github.changmin6362.multistore.feature.authorization.web.response.Perm
 import io.github.changmin6362.multistore.feature.authorization.service.PermissionService;
 import io.github.changmin6362.multistore.feature.authorization.web.request.PermissionCreateRequest;
 import io.github.changmin6362.multistore.feature.authorization.web.request.PermissionUpdateRequest;
-// 컨트롤러 전용 래퍼/DTO 사용 제거: 도메인의 PermissionResponse를 직접 반환
+
 import io.github.changmin6362.multistore.common.web.flags.DeletedResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -69,7 +69,7 @@ public class PermissionController {
      * 권한 단건 조회 (없으면 404)
      */
     @GetMapping("/{permissionId}")
-    public ResponseEntity<ApiResponse> getPermission(@PathVariable Long permissionId) {
+    public ResponseEntity<ApiResponse> getPermission(@PathVariable int permissionId) {
         PermissionResponse permission = permissionService.findById(permissionId);
         if (permission == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
@@ -85,7 +85,7 @@ public class PermissionController {
      * 요청 바디: { "permissionName": "...", "resourceType": "...", "actionType": "...", "permissionDescription": "..." }
      */
     @PutMapping("/{permissionId}")
-    public ResponseEntity<ApiResponse> updatePermission(@PathVariable Long permissionId,
+    public ResponseEntity<ApiResponse> updatePermission(@PathVariable int permissionId,
                                                                              @RequestBody PermissionUpdateRequest body) {
         String permissionName = body != null ? body.permissionName() : null;
         String resourceType = body != null ? body.resourceType() : null;
@@ -110,7 +110,7 @@ public class PermissionController {
      * 권한 삭제 (없으면 404)
      */
     @DeleteMapping("/{permissionId}")
-    public ResponseEntity<ApiResponse> deletePermission(@PathVariable Long permissionId) {
+    public ResponseEntity<ApiResponse> deletePermission(@PathVariable int permissionId) {
         boolean deleted = permissionService.delete(permissionId);
         if (!deleted) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
