+++ backend/src/main/java/io/github/changmin6362/multistore/feature/authorization/web/response/RoleResponse.java
@@ -4,7 +4,7 @@ package io.github.changmin6362.multistore.feature.authorization.web.response;
  * 역할 정보를 담는 응답 전용 레코드
  */
 public record RoleResponse(
-        Long roleId,
+        int roleId,
         String roleName,
         String roleDescription
 ) {
